Move components to dedicated folder
