+++ packages/twenty-front/src/modules/workflow/workflow-diagram/workflow-edges/constants/WorkflowDiagramEdgeOptionsClickOutsideId.ts
