+++ packages/twenty-front/src/modules/workflow/workflow-diagram/workflow-edges/states/workflowHoveredEdgeComponentState.ts
@@ -1,11 +1,11 @@
 import { createComponentState } from '@/ui/utilities/state/component-state/utils/createComponentState';
 import { WorkflowVisualizerComponentInstanceContext } from '@/workflow/workflow-diagram/states/contexts/WorkflowVisualizerComponentInstanceContext';
-import { type WorkflowDiagramEdge } from '@/workflow/workflow-diagram/types/WorkflowDiagramEdge';
+import { type WorkflowDiagramEdge } from '@/workflow/workflow-diagram/workflow-edges/types/WorkflowDiagramEdge';
 
-export const workflowSelectedEdgeComponentState = createComponentState<
+export const workflowHoveredEdgeComponentState = createComponentState<
   WorkflowDiagramEdge | undefined
 >({
-  key: 'workflowSelectedEdgeComponentState',
+  key: 'workflowHoveredEdgeComponentState',
   defaultValue: undefined,
   componentInstanceContext: WorkflowVisualizerComponentInstanceContext,
 });
